requirements.txt
jira
PyGithub
python-dotenv
streamlit
